requirements.txt
diffusers==0.31.0
numpy==1.26.4
Pillow==11.0.0
PyQt6==6.7.1
PyQt6_sip==13.8.0
torch==2.2.2
torchvision==0.17.2
accelerate==1.0.1
transformers==4.46.1